requirements.txt
streamlit
streamlit_chat
openai
python-dotenv
watchdog
pydantic
pyyaml
shyaml